Move utils from advice-animal-macro.jsx to utils.jsx
